requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0  # Para rodar localmente com FastAPI
python-multipart>=0.0.5    # Para upload de arquivos com FastAPI
google-generativeai>=0.5.0 # SDK do Gemini
python-dotenv>=1.0.0       # Para carregar .env em desenvolvimento local
# appdirs                  # Não é mais estritamente necessário para config web
# PyQt5                    # Removido, não usado na versão web
# qtawesome                # Removido
# darkdetect               # Removido

# Adicione outras dependências que seu projeto possa ter.
